Added UIText Field to Upper TableView
The UITableViewCell proved difficult to get the right behaviour with auto layout.  A needed flag to AutomaticallyAdjustsScrollViewInsets was needed to set to false to stop the scroll from wanting to go under the navigation bar which pushed the cell down further then the height allowed.  This is now fixed and a accessory view button is added.  More experimentation with the TopLayoutGuide is needed.