Dynamic Data Table Functionality
Added the ability for the bottom table to have a predefined setup without a massive GetCell Method.  There is now a set structure of cells that makes proper use of cell dequeing and sets up the bottom table view.  This structure facilitiates the passage of data into the table view and because of this sould be easily testable.